requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
yt-dlp>=2025.6.9
pydantic-settings==2.1.0
httpx==0.27.0
smbprotocol==1.13.0
rich==13.7.0
google-api-python-client==2.111.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pytz==2024.1
